pyproject.toml
[project]
name = "qsis-renew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "numpy>=2.4.2",
    "paho-mqtt>=2.1.0",
    "phidget22>=1.24.20260127",
    "plotly>=6.5.2",
    "scipy>=1.17.1",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
